Fix code-style and float16 precision issue in _convolution test
- Remove unused `import os`
- Remove `.to(dtype)` on reference output to avoid float16 precision loss
  that caused CI test failure (greatest absolute diff 0.00097 > 0.0001 allowed)

+++ tests/test__convolution.py
@@ -1,5 +1,3 @@
-import os
-
 import pytest
 import torch
 
@@ -64,7 +62,7 @@ def test_accuracy__convolution(
         False,  # deterministic
         True,  # cudnn_enabled
         True,  # allow_tf32
-    ).to(dtype)
+    )
 
     with flag_gems.use_gems():
         res_out = torch._convolution(
@@ -126,7 +124,7 @@ def test_accuracy__convolution_1d(monkeypatch, shape, kernel, stride, padding, d
         False,  # deterministic
         True,  # cudnn_enabled
         True,  # allow_tf32
-    ).to(dtype)
+    )
 
     with flag_gems.use_gems():
         res_out = torch._convolution(
@@ -191,7 +189,7 @@ def test_accuracy__convolution_3d(
         False,  # deterministic
         True,  # cudnn_enabled
         True,  # allow_tf32
-    ).to(dtype)
+    )
 
     with flag_gems.use_gems():
         res_out = torch._convolution(
